Renamed artifactId and packages, updated the added temp reading error field for failing tests
