+++ IoT Devices/Pump Controller/pump-controller/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>quintin.raspberry</groupId>
+    <groupId>piandarduinoguy.raspberry</groupId>
     <artifactId>pump-controller</artifactId>
     <version>DEV-SNAPSHOT</version>
 
@@ -76,7 +76,7 @@
                         <configuration>
                             <classifier>spring-boot</classifier>
                             <mainClass>
-                                quintin.raspberrypi.pump_controller.PumpControllerApplication
+                                piandarduinoguy.raspberrypi.pump_controller.PumpControllerApplication
                             </mainClass>
                         </configuration>
                     </execution>
