+++ IoT Devices/Pump Controller/pump-controller/src/main/java/piandarduinoguy/raspberrypi/pump_controller/PumpControllerApplication.java
@@ -1,0 +1,12 @@
+package piandarduinoguy.raspberrypi.pump_controller;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PumpControllerApplication {
+
+    public static void main(String... args) {
+        SpringApplication.run(PumpControllerApplication.class, args);
+    }
+}
