+++ IoT Devices/Pump Controller/pump-controller/src/main/java/piandarduinoguy/raspberrypi/pump_controller/data/OverrideStatus.java
@@ -1,0 +1,5 @@
+package piandarduinoguy.raspberrypi.pump_controller.data;
+
+public enum OverrideStatus {
+    NONE, PUMP_ON, PUMP_OFF, NOT_SET
+}
