+++ IoT Devices/Pump Controller/pump-controller/src/main/java/piandarduinoguy/raspberrypi/pump_controller/data/Problem.java
@@ -1,0 +1,19 @@
+package piandarduinoguy.raspberrypi.pump_controller.data;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class Problem {
+    private String title;
+    private Integer status;
+    private String detail;
+}
