+++ IoT Devices/Pump Controller/pump-controller/src/main/java/piandarduinoguy/raspberrypi/pump_controller/domain/PumpState.java
@@ -1,0 +1,12 @@
+package piandarduinoguy.raspberrypi.pump_controller.domain;
+
+public enum PumpState {
+    ON("ON"),
+    OFF("OFF");
+
+    private String pumpState;
+
+    PumpState(String pumpState){
+        this.pumpState = pumpState;
+    }
+}
