+++ IoT Devices/Pump Controller/pump-controller/src/main/java/piandarduinoguy/raspberrypi/pump_controller/exception/PumpControllerException.java
@@ -1,0 +1,7 @@
+package piandarduinoguy.raspberrypi.pump_controller.exception;
+
+public class PumpControllerException extends RuntimeException {
+    public PumpControllerException(String message) {
+        super(message);
+    }
+}
