+++ IoT Devices/Pump Controller/pump-controller/src/main/java/piandarduinoguy/raspberrypi/pump_controller/observable/NewAmbientTempReadingObservable.java
@@ -1,0 +1,18 @@
+package piandarduinoguy.raspberrypi.pump_controller.observable;
+
+import lombok.Getter;
+import org.springframework.stereotype.Component;
+
+import java.util.Observable;
+
+@Component
+@Getter
+public class NewAmbientTempReadingObservable extends Observable {
+    private double tempReading;
+
+    public void setTemp(double newTempReading){
+        this.tempReading = newTempReading;
+        this.setChanged();
+        this.notifyObservers(tempReading);
+    }
+}
