--- IoT Devices/Pump Controller/pump-controller/src/main/java/quintin/raspberrypi/pump_controller/data/OverrideStatus.java
@@ -1,5 +1,0 @@
-package quintin.raspberrypi.pump_controller.data;
-
-public enum OverrideStatus {
-    NONE, PUMP_ON, PUMP_OFF, NOT_SET
-}
