--- IoT Devices/Pump Controller/pump-controller/src/main/java/quintin/raspberrypi/pump_controller/data/Problem.java
@@ -1,19 +1,0 @@
-package quintin.raspberrypi.pump_controller.data;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class Problem {
-    private String title;
-    private Integer status;
-    private String detail;
-}
