--- IoT Devices/Pump Controller/pump-controller/src/main/java/quintin/raspberrypi/pump_controller/data/PumpConfig.java
@@ -1,19 +1,0 @@
-package quintin.raspberrypi.pump_controller.data;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-@Component
-public class PumpConfig {
-    private double turnOnTemp;
-    private OverrideStatus overrideStatus = OverrideStatus.NOT_SET;
-}
