--- IoT Devices/Pump Controller/pump-controller/src/main/java/quintin/raspberrypi/pump_controller/domain/PumpState.java
@@ -1,12 +1,0 @@
-package quintin.raspberrypi.pump_controller.domain;
-
-public enum PumpState {
-    ON("ON"),
-    OFF("OFF");
-
-    private String pumpState;
-
-    PumpState(String pumpState){
-        this.pumpState = pumpState;
-    }
-}
