--- IoT Devices/Pump Controller/pump-controller/src/main/java/quintin/raspberrypi/pump_controller/exception/PumpControllerException.java
@@ -1,7 +1,0 @@
-package quintin.raspberrypi.pump_controller.exception;
-
-public class PumpControllerException extends RuntimeException {
-    public PumpControllerException(String message) {
-        super(message);
-    }
-}
