--- IoT Devices/Pump Controller/pump-controller/src/main/java/quintin/raspberrypi/pump_controller/observable/PumpOverrideStatusObservable.java
@@ -1,19 +1,0 @@
-package quintin.raspberrypi.pump_controller.observable;
-
-import lombok.Getter;
-import org.springframework.stereotype.Component;
-import quintin.raspberrypi.pump_controller.data.OverrideStatus;
-
-import java.util.Observable;
-
-@Component
-@Getter
-public class PumpOverrideStatusObservable extends Observable {
-    private OverrideStatus overrideStatus;
-
-    public void setOverrideStatus(OverrideStatus overrideStatus) {
-        this.overrideStatus = overrideStatus;
-        this.setChanged();
-        this.notifyObservers(overrideStatus);
-    }
-}
