+++ IoT Devices/Pump Controller/pump-controller/src/test/resources/application-test.properties
@@ -1,3 +1,5 @@
 automatic-pump-toggler.readings-before-check=5
 ambient-temp-reader.scheduler.delay-seconds=5 
 ambient-temp-reader.scheduler.rate-seconds=5
+# error of thermistor measured temperature in degrees celsius
+temp-reading.error=2
