+++ IoT Devices/Pump Controller/pump-controller/src/test/resources/application-valid-python-script-location.properties
@@ -2,3 +2,5 @@ mcp3002.python-script.location=/scripts/mcp3002_adc_value.py
 automatic-pump-toggler.readings-before-check=5
 ambient-temp-reader.scheduler.delay-seconds=5 
 ambient-temp-reader.scheduler.rate-seconds=5
+# error of thermistor measured temperature in degrees celsius
+temp-reading.error=2
